Rearranging logic to deal with timing bugs.
Previously there was a bug that could be reproduced by selecting typing
something very quickly, selecting all of the text and then deleting it.
A few moments later, the previous value would magically re-appear in the
text box.

This occured, because we were checking for minTermLength before
clearing the timer.  So the method was returning false b/c of query
length before the timer could be cleared.

Theoretically if the input text value has changed from prevValue, the
previous query request is invalid, so it should be cleared.  I moved
this logic nearer to the top of the function and checked for the minTermLength
further down, after the timer has already been cleared.